refactor(api-gen): remove `constants` and `helpers` directories
This moves the last of the code from these catcj-all directories to more
specific homes.
